Unpatching serdi changes
Serdi has been replaced again with rapper due to serdi just not being correct in the utilisation of prefixes.
Even when prefixes are added to the turtle it interprets them incorrectly. In addition it is not able
to properly reconstruct rdf lists. For this reason we'll have to work with rapper and add it as a dependency.
Ultimately we need a better solution here. It's unlikely that people will want to edit JSON-LD, so we may need better turtle processing.
The serd library looks to be more flexible than the command line tool as well, so it's possible we can
implement bindings that will improve the situation.