electron: get rid of manual process killing
This should fix #205 and should be fixed by running
serve with the --interactive flag.

If serve is not run with interactive flag it does not respond
well to ANY signals that the parent process sends. I think that
serve is fundamentally broken without --interactive. It is good
that it runs in a deamon-like mode but is totally unresponsive in
a way in which normal daemon-like applications shouldn't.