Refactor and optimize list db
Increased performance further. If the account requesting the list
is the super_user/admin, there is no need to get the user object
and query the organization of *every* database.

However, even when a specific user wants a list of their DBs
it is also faster now. Since the name of the database is already
in the user object itself, it does not need to be queried.
This saves addiontal time and also removes the need for the
ground check of the DB URI.

Time went from:
./terminusdb list  6.30s user 1.87s system 103% cpu 7.912 total

To:
./terminusdb list  2.37s user 1.21s system 107% cpu 3.316 total