Fix organization filter (#313)
The organization filter was too strict. If the organization
name contained one of the excluded names, it would already
forbid this organization too exist. Even a organization name
called "soundboard" would be forbidden by the existing filter.

Now it only checks whether the whole name is the forbidden word.
The exception to this rule is the | (pipe) character which
should always be forbidden.