Add ability to authenticate using a custom username header
Some systems can be setup in such a way to offload user
authentication. For instance, you could set up TerminusDB
in a protected environment and use oauth2-proxy to secure it.

Therefore we now allow an insecure mode to be set in conjunction
with a custom user header to authenticate.

The ENV variables required for this functionality are:

TERMINUSDB_INSECURE_MODE which has to be set to a value
and TERMINUSDB_FORWARDED_USER_HEADER with the name of the
HTTP header which contains the username.