Add simple script to lint and execute it in CI
Very simple script to check for missing autoloading imports.
It does the following step by steps:

1. It downloads the SWI Prolog help page and gets all the
   built-in library predicates
2. It traverses every prolog file
3. While it reads a file, it will check whether a predicate
   exists without a certain library import

It will skip two libraries: help and statistics. We will
probably never use them and they lead to false positives
that are hard to get away it. I know that this is a very
dirty hack that could probably be better implemented in
Prolog itself, but I do think that this issue is so
important for our code quality, that a more-or-less
hacky linter is better than no linter at all.