capabilities: fix user search and test with header auth
If the user is authenticated with http basic, the full URI
is fetched from the database instead of the prefixed
result. However, when you authenticate with JWT or a header,
it won't. This led to bugs in different API endpoints.

It also runs the tests with the insecure header because it
needs to be tested as well. I agree that we'll need something
better like Sean suggested in issue #941 but I do think
that this is necessary in the meantime.