Remove JWT support
We rely on a Prolog JWT library which is not very well
maintained anymore. API gateways like Kong provide
better JWT support and should be used instead of
doing the same verification twice.

The insecure user header can be used instead to configure
a proxy to check for the JWTs.