Use 'unsafe' document insert for full_replace
After #1095, we started using the 'normal' document insert for
full replace, but it checked for the existence of the document ID on
insert, which meant it failed when inserting the same document again.

Now, we use an 'unsafe' document insert for full replace, which doesn't
check for the ID. We already nuke all documents, so this is fine.

This also adds integration tests which were missing from #1095.

Fixes #1112