Use random storage subdirectory for CLI tests
Currently, the CLI integration tests force initialize the store, which
removes any existing default store (`./storage/db`). This may not be
desirable if you are currently using that store for development.

However, sometimes, the tests fail or are killed, and it is useful to
remove any existing store when re-running the tests.

This change uses the `TERMINUSDB_SERVER_DB_PATH` environment variable to
set the store to a random subdirectory of `./storage`. We can therefore
both avoid removing the default store and remove any previously created
random subdirectory used for tests.

Fixes #1085.