Push pull clone (#1124)
* Add command line token connection

* Fix the way the pack/unpack urls are built to work with terminusx

* fix lint issues

* Add token auth to CLI

* Fix url formation for TUS

* Add test for unknown property (#1100)

This is part of #1030.

* Fix CLI query message option (#1116)

* Use random storage subdirectory for CLI tests (#1115)

Currently, the CLI integration tests force initialize the store, which
removes any existing default store (`./storage/db`). This may not be
desirable if you are currently using that store for development.

However, sometimes, the tests fail or are killed, and it is useful to
remove any existing store when re-running the tests.

This change uses the `TERMINUSDB_SERVER_DB_PATH` environment variable to
set the store to a random subdirectory of `./storage`. We can therefore
both avoid removing the default store and remove any previously created
random subdirectory used for tests.

Fixes #1085.

* Initial basic document insert and get in CLI (#1120)

Add `doc insert` and `doc get` subcommands to the TerminusDB CLI.

Initially, these will allow us to test the CLI more easily. The input is
`stdin`, and the output is `stdout`. The options available are limited. In the
future, we will add more functionality to these subcommands.

* Upgrade tus library to latest version in the docker (#1121)

* Add command line token connection

* Add token auth to CLI

* Remove double create auth

* Added insecure user header cut

* Adding tus 0.0.7 api base options

* Adding correct tus

* Adding comma

* Adding headers

Co-authored-by: Matthijs van Otterdijk <matthijs@terminusdb.com>
Co-authored-by: Sean Leather <sean@terminusdb.com>