swipl --on-error=halt --on-warning=halt
Use these flags with `swipl` in the `Makefile` to tell it to halt with a
non-zero exit code when the first error or warning occurs.

These flags have been available since SWI-Prolog 8.4. We are currently
on 8.4.2.

See for usage:
<https://swi-prolog.discourse.group/t/changes-in-error-handling/3930>

This also appears to fix #873, probably because I removed the `grep`
commands in the pipes. The next release of SWI-Prolog will not need
these thanks to: <https://github.com/SWI-Prolog/swipl/pull/22>