Add benchmarking on dedicated machine (#1491)
* ci: try to publish js benchmarks to github pages



Benchmarks have been enabled to run on a remote machine. The results are published on the gh-pages branch as can be seen here with the few test runs I did: https://github.com/terminusdb/terminusdb/tree/gh-pages

The following benchmarks are executed on the remote machine:

    The k6 benchmarks https://github.com/terminusdb-labs/terminusdb-http-perf
    The JS benchmarks in tests/bench
    Lego data is ingested from https://github.com/terminusdb-labs/terminus-cms/ which is a relative big JSON. It is measured by using GNU Time

The server runs a relatively simple Go program that executes those benchmarks and saves those on the machine. It is started from GH Actions with SSH and subsequently fetched by using SCP.

The Go benchmark program can be found here: https://github.com/terminusdb-labs/terminusdb-benchmark-runner

The benchmark workflow is configured to only run on merges to main because it takes a long time and would otherwise slow down development a lot. It also doesn't run concurrently because that would result in less reliable benchmarks.