Remove built-in server welcome message (#1696)
This message is not needed as we already print our own welcome
message, which includes on which port it is running.

It is also printed regardless of whether you are setting the log
format to JSON which is very annoying as you get JSON parsing
errors that way. Removing that green welcome message solves
the problem and I do think that our own welcome mesage is
clear enough.