ci: add major tag
If a new release is tagged, we also push a major tag now without
containing the full version. For instance, when someone pushes
a new tag called v11.0.4, it will update the v11 tag with this
new image.

The reason for this is that it is easier for users to track minor
releases this way.